removed TextArea1 from directory and replaced with textArea
